\name{Internals}
\alias{Internals} 
\alias{getC}
\alias{getCList}
\alias{getS}
\alias{barplot3}
\alias{normdat}
\alias{getWSList}
\alias{getWSListAllTimes}
\title{MCR-ALS functions used internally}
\description{MCR-ALS functions used internally}
\author{Katharine M. Mullen}
\seealso{\code{\link{als}}} 
\keyword{internal}

